Module bevy_xpbd_2d::components

Components used for rigid bodies, colliders and mass properties.

Structs

AngularDamping

Automatically slows down a dynamic rigid body, decreasing it’s angular velocity each frame. This can be used to simulate air resistance.

AngularVelocity

The angular velocity of a body in radians. Positive values will result in counterclockwise rotation.

CenterOfMass

The local center of mass of a body.

Collider

A collider used for collision detection.

ColliderAabb

The Axis-Aligned Bounding Box of a collider.

ColliderMassProperties

The mass properties derived from a given collider shape and density.

CollidingEntities

Contains the entities that are colliding with an entity.

CollisionLayers

Defines the collision layers of a collider using groups and masks.

ExternalForce

An external force applied to a dynamic rigid body during the integration step.

ExternalTorque

An external torque applied to a dynamic rigid body during the integration step.

Friction

Controls how strongly the material of an entity prevents relative tangential movement at contact points.

GravityScale

Controls how gravity affects a specific rigid body.

Inertia

The moment of inertia of a body. This represents the torque needed for a desired angular acceleration.

InverseInertia

The inverse moment of inertia of the body. This represents the inverse of the torque needed for a desired angular acceleration.

InverseMass

The inverse mass of a body.

LinearDamping

Automatically slows down a dynamic rigid body, decreasing it’s linear velocity each frame. This can be used to simulate air resistance.

LinearVelocity

The linear velocity of a body.

LockedAxes

A component that specifies which translational and rotational axes of a rigid body are locked.

Mass

The mass of a body.

MassPropertiesBundle

A bundle containing mass properties.

Position

The position of a body.

PreviousPosition

The previous position of a body.

PreviousRotation

The previous rotation of a body. See Rotation.

Restitution

Controls how elastic or bouncy an entity is when colliding with other entities.

RigidBodyQuery

A WorldQuery to make querying and modifying rigid bodies more convenient.

RigidBodyQueryItem

Automatically generated WorldQuery item type for RigidBodyQuery, returned when iterating over query results.

RigidBodyQueryReadOnly

Automatically generated WorldQuery type for a read-only variant of RigidBodyQuery.

RigidBodyQueryReadOnlyItem

Automatically generated WorldQuery item type for RigidBodyQueryReadOnly, returned when iterating over query results.

Rotation

The rotation of a body.

Sensor

A component that marks a Collider as a sensor collider.

Sleeping

Indicates that a body is not simulated by the physics engine until woken up again. This is done to improve performance and to help prevent small jitter that is typically present in collisions.

SleepingDisabled

Indicates that the body can not be deactivated by the physics engine. See Sleeping for information about sleeping.

TimeSleeping

How long the velocity of the body has been below the SleepingThreshold, i.e. how long the body has been able to sleep.

Enums

CoefficientCombine

Determines how coefficients are combined. The default is Average.

RigidBody

A non-deformable body used for the simulation of most physics objects.

Traits

PhysicsLayer

A layer used for determining which entities should interact with each other. Physics layers are used heavily by CollisionLayers.