Struct bevy_xpbd_2d::components::Collider

pub struct Collider(/* private fields */);

A collider used for collision detection.

By default, colliders generate collision events and cause a collision response for rigid bodies. If you only want collision events, you can add a Sensor component.

Creation

Collider has tons of methods for creating colliders of various shapes. For example, to add a ball collider to a rigid body, use Collider::ball:

use bevy::prelude::*;
use bevy_xpbd_3d::prelude::*;

fn setup(mut commands: Commands) {
    commands.spawn((RigidBody::Dynamic, Collider::ball(0.5)));
}

In addition, Bevy XPBD will automatically add some other components, like the following:

ColliderAabb

CollidingEntities

ColliderMassProperties

Collision layers

You can use collsion layers to configure which entities can collide with each other.

See CollisionLayers for more information and examples.

Collision events

There are currently three different collision events: Collision, CollisionStarted and CollisionEnded. You can listen to these events as you normally would.

For example, you could read contacts like this:

use bevy::prelude::*;
use bevy_xpbd_3d::prelude::*;

fn my_system(mut collision_event_reader: EventReader<Collision>) {
    for Collision(contact) in collision_event_reader.iter() {
        println!("{:?} and {:?} are colliding", contact.entity1, contact.entity2);
    }
}

Advanced usage

Internally, Collider uses the shapes provided by parry. If you want to create a collider using these shapes, you can simply use Collider::from(SharedShape::some_method()).

To get a reference to the internal SharedShape, you can use the get_shape method.

Implementations

impl Collider

pub fn get_shape(&self) -> &SharedShape

Gets the raw shape of the collider. The shapes are provided by parry.

pub fn compute_aabb(&self, position: Vector, rotation: Scalar) -> ColliderAabb

Computes the Axis-Aligned Bounding Box of the collider.

pub fn compound( shapes: Vec<(impl Into<Position>, impl Into<Rotation>, impl Into<Collider>)> ) -> Self

Creates a collider with a compound shape defined by a given vector of colliders with a position and a rotation.

Especially for dynamic rigid bodies, compound shape colliders should be preferred over triangle meshes and polylines, because convex shapes typically provide more reliable results.

If you want to create a compound shape from a 3D triangle mesh or 2D polyline, consider using the Collider::convex_decomposition method.

pub fn ball(radius: Scalar) -> Self

Creates a collider with a ball shape defined by its radius.

pub fn cuboid(x_length: Scalar, y_length: Scalar) -> Self

Creates a collider with a cuboid shape defined by its extents.

pub fn capsule(height: Scalar, radius: Scalar) -> Self

Creates a collider with a capsule shape defined by its height along the Y axis and its radius.

pub fn capsule_endpoints(a: Vector, b: Vector, radius: Scalar) -> Self

Creates a collider with a capsule shape defined by its end points a and b and its radius.

pub fn halfspace(outward_normal: Vector) -> Self

Creates a collider with a half-space shape defined by the outward normal of its planar boundary.

pub fn segment(a: Vector, b: Vector) -> Self

Creates a collider with a segment shape defined by its endpoints a and b.

pub fn triangle(a: Vector, b: Vector, c: Vector) -> Self

Creates a collider with a triangle shape defined by its points a, b and c.

pub fn polyline(vertices: Vec<Vector>, indices: Option<Vec<[u32; 2]>>) -> Self

Creates a collider with a polyline shape defined by its vertices and optionally an index buffer.

pub fn trimesh(vertices: Vec<Vector>, indices: Vec<[u32; 3]>) -> Self

Creates a collider with a triangle mesh shape defined by its vertex and index buffers.

pub fn convex_decomposition( vertices: Vec<Vector>, indices: Vec<[u32; 2]> ) -> Self

Creates a collider shape with a compound shape obtained from the decomposition of a given polyline defined by its vertex and index buffers.

pub fn convex_hull(points: Vec<Vector>) -> Option<Self>

Creates a collider with a convex polygon shape obtained after computing the convex hull of the given points.

pub fn heightfield(heights: Vec<Scalar>, scale: Scalar) -> Self

Creates a collider with a heightfield shape.

A 2D heightfield is a segment along the X axis, subdivided at regular intervals.

heights is a vector indicating the altitude of each subdivision point, and scale is a scalar value indicating the length of each subdivided segment along the X axis.