Module bevy_xpbd_2d::plugins

The building blocks of Bevy XPBD.

Bevy XPBD consists of multiple different plugins, each with their own responsibilities. These plugins are grouped into the PhysicsPlugins plugin group, which allows you to easily configure and disable any of the existing plugins.

This means that users can simply disable existing functionality and replace it with specialized solutions, while keeping the rest of the engine’s features. This can be important in many games and applications that require special optimizations or features.

Currently, the only plugin that is truly necessary is the PhysicsSetupPlugin that initializes the engine’s resources, schedules and sets, and even that could be implemented manually if you wanted a custom scheduler for example.

See also

All of the current plugins and their responsibilities

Creating custom plugins

PhysicsSet

PhysicsSchedule and PhysicsStepSet

SubstepSchedule and SubstepSet

Re-exports

pub use broad_phase::BroadPhasePlugin;

pub use integrator::IntegratorPlugin;

pub use prepare::PreparePlugin;

pub use sleeping::SleepingPlugin;

pub use solver::solve_constraint;

pub use solver::SolverPlugin;

pub use sync::SyncPlugin;

pub use setup::*;

pub use spatial_query::*;

Modules

broad_phase

Collects pairs of potentially colliding entities into BroadCollisionPairs using AABB intersection checks.

integrator

Integrates Newton’s 2nd law of motion, applying forces and moving entities according to their velocities.

prepare

Runs systems at the start of each physics frame; initializes rigid bodies and colliders and updates components.

setup

Sets up the physics engine by initializing the necessary schedules, sets and resources.

sleeping

Controls when bodies should be deactivated and marked as Sleeping to improve performance.

solver

Solves positional and angular constraints, updates velocities and solves velocity constraints (dynamic friction, restitution and joint damping).

spatial_query

Spatial queries are a way to get information about the environment. They perform geometric queries on colliders and retrieve data about intersections.

sync

Responsible for synchronizing physics components with other data, like writing Position and Rotation components to Transforms.

Structs

PhysicsPlugins

A plugin group containing all of Bevy XPBD’s plugins.