Module bevy_xpbd_2d::plugins::collision

Collision detection is used to detect and compute intersections between Colliders.

In bevy_xpbd, collision detection is split into three plugins:

BroadPhasePlugin: Collects pairs of potentially colliding entities into BroadCollisionPairs.

NarrowPhasePlugin: Computes contacts for broad phase collision pairs and adds them to Collisions.

ContactReportingPlugin (optional): Sends collision events and updates CollidingEntities based on Collisions.

Spatial queries are handled by the SpatialQueryPlugin.

You can also find several utility methods for computing contacts in contact_query.

Modules

broad_phase

Collects pairs of potentially colliding entities into BroadCollisionPairs using AABB intersection checks.

contact_query

Contact queries compute information about contacts between two Colliders.

contact_reporting

Sends collision events and updates CollidingEntities.

narrow_phase

Computes contacts between entities.

Structs

Collisions

A resource that stores all collision pairs.

ContactData

Data related to a contact between two bodies.

ContactManifold

A contact manifold between two colliders, containing a set of contact points. Each contact in a manifold shares the same contact normal.

Contacts

All contacts between two colliders.