Type Alias bevy_xpbd_2d::plugins::collision::contact_query::TimeOfImpactStatus

pub type TimeOfImpactStatus = TOIStatus;

The way the time of impact computation was terminated.

Aliased Type

enum TimeOfImpactStatus {
    OutOfIterations,
    Converged,
    Failed,
    Penetrating,
}

Variants

OutOfIterations

The TOI algorithm ran out of iterations before achieving convergence.

The content of the TOI will still be a conservative approximation of the actual result so it is often fine to interpret this case as a success.

Converged

The TOI algorithm converged successfully.

Failed

Something went wrong during the TOI computation, likely due to numerical instabilities.

The content of the TOI will still be a conservative approximation of the actual result so it is often fine to interpret this case as a success.

Penetrating

The two shape already overlap at the time 0.

The witness points and normals provided by the TOI will have undefined values.