Trait bevy_xpbd_2d::plugins::debug::PhysicsGizmoExt

pub trait PhysicsGizmoExt {
    // Required methods
    fn draw_line(&mut self, a: Vector, b: Vector, color: Color);
    fn draw_line_strip(
        &mut self,
        points: Vec<Vector>,
        position: impl Into<Position>,
        rotation: impl Into<Rotation>,
        closed: bool,
        color: Color
    );
    fn draw_polyline(
        &mut self,
        vertices: &[Vector],
        indices: &[[u32; 2]],
        position: impl Into<Position>,
        rotation: impl Into<Rotation>,
        color: Color
    );
    fn draw_arrow(
        &mut self,
        a: Vector,
        b: Vector,
        head_length: Scalar,
        color: Color
    );
    fn draw_collider(
        &mut self,
        collider: &Collider,
        position: impl Into<Position>,
        rotation: impl Into<Rotation>,
        color: Color
    );
    fn draw_raycast(
        &mut self,
        origin: Vector,
        direction: Direction2d,
        max_time_of_impact: Scalar,
        hits: &[RayHitData],
        ray_color: Color,
        point_color: Color,
        normal_color: Color
    );
    fn draw_shapecast(
        &mut self,
        shape: &Collider,
        origin: Vector,
        shape_rotation: impl Into<Rotation>,
        direction: Direction2d,
        max_time_of_impact: Scalar,
        hits: &[ShapeHitData],
        ray_color: Color,
        shape_color: Color,
        point_color: Color,
        normal_color: Color
    );
}

An extension trait for Gizmos<PhysicsGizmo>.

Required Methods

fn draw_line(&mut self, a: Vector, b: Vector, color: Color)

Draws a line from a to b.

fn draw_line_strip( &mut self, points: Vec<Vector>, position: impl Into<Position>, rotation: impl Into<Rotation>, closed: bool, color: Color )

Draws lines between a list of points.

fn draw_polyline( &mut self, vertices: &[Vector], indices: &[[u32; 2]], position: impl Into<Position>, rotation: impl Into<Rotation>, color: Color )

Draws a polyline based on the given vertex and index buffers.

fn draw_arrow( &mut self, a: Vector, b: Vector, head_length: Scalar, color: Color )

Draws an arrow from a to b with an arrowhead that has a length of head_length.

fn draw_collider( &mut self, collider: &Collider, position: impl Into<Position>, rotation: impl Into<Rotation>, color: Color )

Draws a Collider shape.

fn draw_raycast( &mut self, origin: Vector, direction: Direction2d, max_time_of_impact: Scalar, hits: &[RayHitData], ray_color: Color, point_color: Color, normal_color: Color )

Draws the results of a raycast.

fn draw_shapecast( &mut self, shape: &Collider, origin: Vector, shape_rotation: impl Into<Rotation>, direction: Direction2d, max_time_of_impact: Scalar, hits: &[ShapeHitData], ray_color: Color, shape_color: Color, point_color: Color, normal_color: Color )

Draws the results of a shapecast.

Object Safety

This trait is not object safe.