Module bevy_xpbd_2d::plugins::prepare

Runs systems at the start of each physics frame. Initializes rigid bodies and updates components.

See PreparePlugin.

Structs

PrepareConfig

Configures what is initialized by the PreparePlugin and how.

PreparePlugin

Runs systems at the start of each physics frame. Initializes rigid bodies and updates components.

Enums

PrepareSet

Systems sets for initializing and syncing missing components. You can use these to schedule your own initialization systems without having to worry about implementation details.

Functions

init_transforms

Initializes Transform based on Position and Rotation or vice versa when a component of the given type is inserted.

update_mass_properties

Updates each body’s InverseMass and InverseInertia whenever Mass or Inertia are changed.