Crate bevy_xpbd_2d

Bevy XPBD

Bevy XPBD is a 2D and 3D physics engine based on Extended Position Based Dynamics (XPBD) for the Bevy game engine.

Design

Below are some of the core design principles used in Bevy XPBD.

Made with Bevy, for Bevy. No wrappers around existing engines.

Provide an ergonomic and familiar API. Ergonomics is key for a good experience.

Utilize the ECS as much as possible. The engine should feel like a part of Bevy, and it shouldn’t need to maintain a separate physics world.

Use a highly modular plugin architecture. Users should be able to replace parts of the engine with their own implementations.

Have good documentation. A physics engine is pointless if you don’t know how to use it.

Features

Below are some of the features of Bevy XPBD.

Dynamic, kinematic and static rigid bodies

Colliders powered by parry

Collision events: Collision, CollisionStarted, CollisionEnded

Access to colliding entities

Sensor colliders

Collision layers

Material properties like restitution and friction

Linear damping and angular damping for simulating drag

Gravity and gravity scale

External forces and torque

Locking translational and rotational axes

Joints

Built-in constraints and support for custom constraints

Spatial queries

Ray casting

Shape casting

Point projection

Intersection tests

Automatically deactivating bodies with sleeping

Configurable timesteps and substepping

f32/f64 precision (f32 by default)

Getting started

This short guide should help you get started with Bevy XPBD.

Add the dependency

First, add bevy_xpbd_2d or bevy_xpbd_3d to your dependencies in Cargo.toml:

# For 2D applications:
[dependencies]
bevy_xpbd_2d = "0.2"

# For 3D applications:
[dependencies]
bevy_xpbd_3d = "0.2"

# If you want to use the most up-to-date version, you can follow the main branch:
[dependencies]
bevy_xpbd_3d = { git = "https://github.com/Jondolf/bevy_xpbd", branch = "main" }

By default, Bevy XPBD uses f32 numbers. If you encounter instability or use a large number of substeps, you might want to use f64 instead. You can change these kinds of features by disabling the default features and manually specifying the feature flags you want:

[dependencies]
# Add 3D Bevy XPBD with double-precision floating point numbers
bevy_xpbd_3d = { version = "0.2", default-features = false, features = ["3d", "f64"] }

Feature flags

Default features: 2d/3d, f32 and collider-from-mesh

2d enables simulation on the x and y axes. Enabled by default for bevy_xpbd_2d. Incompatible with 3d.

3d enables simulation on the x, y and z axes. Enabled by default for bevy_xpbd_3d. Incompatible with 2d.

f32 enables using f32 numbers. Incompatible with f64.

f64 enables using f64 numbers. Recommended when encountering stability problems, especially with small timesteps. Incompatible with f32.

debug-plugin enables the PhysicsDebugPlugin used for rendering physics objects and events like AABBs and contacts.

collider-from-mesh allows you to create colliders from Bevy meshes. Enables bevy_render.

simd enables SIMD optimizations.

enhanced-determinism enables increased determinism. (Note: cross-platform determinism doesn’t work yet, even with this feature enabled)

Install the plugin

Bevy XPBD is designed to be very modular. It is built from many different plugins that manage different parts of the engine. These plugins can be easily initialized and configured through the PhysicsPlugins plugin group.

use bevy::prelude::*;
use bevy_xpbd_3d::prelude::*;

fn main() {
    App::new()
        .add_plugins((DefaultPlugins, PhysicsPlugins::default()))
        // ...your other plugins, systems and resources
        .run();
}

Now you can use all of Bevy XPBD’s components and resources to build whatever you want!

For example, adding a rigid body with a collider is as simple as spawning an entity with the RigidBody and Collider components:

fn setup(mut commands: Commands) {
    commands.spawn((RigidBody::Dynamic, Collider::ball(0.5)));
}

To learn more about using Bevy XPBD, consider taking a look at the official examples and how to accomplish some common tasks.

To learn more about the structure of the engine, consider taking a look at the plugins and what XPBD actually is.

Examples

You can find 2D examples in crates/bevy_xpbd_2d/examples and 3D examples in crates/bevy_xpbd_3d/examples.

You can run the examples with default features as you normally would. For example, running the cubes example looks like this:

cargo run --example cubes

Note that the examples support both f32 and f64 precisions, so the code contains some feature-dependent types like Scalar and Vector. In actual usage these are not needed, so you can just use f32 or f64 types depending on the features you have chosen.

By default the examples use f32, so if you want to run the f64 versions, you need to disable the default features and manually choose the dimension and precision:

cargo run --example cubes --no-default-features --features "3d f64"

Common tasks

Create a rigid body

Define mass properties

Add a collider

Listen to collision events

Define collision layers

Configure restitution

Configure friction

Configure gravity

Apply external forces

Apply external torque

Lock translational and rotational axes

Use joints

Perform spatial queries

Ray casting

Shape casting

Point projection

Intersection tests

Configure the physics timestep

Configure the substep count

Configure the schedule for running physics

Usage on servers

Create custom constraints

Replace built-in plugins with custom plugins

Frequently asked questions

How does Bevy XPBD compare to Rapier and bevy_rapier?

Rapier is the biggest and most used physics engine in the Rust ecosystem, and it is currently the most mature and feature-rich option.

bevy_rapier is a great physics integration for Bevy, but it does have several problems:

It has to maintain a separate physics world and synchronize a ton of data with Bevy each frame

The source code is difficult to inspect, as the vast majority of it is glue code and wrappers for Bevy

It has poor docs.rs documentation, and the documentation on rapier.rs is often outdated and missing features

It is hard to extend as it’s not very modular or composable in design

Overall, it doesn’t have a native ECS-like feel outside of its public API

Bevy XPBD on the other hand is built for Bevy with Bevy, and it uses the ECS for both the internals and the public API. This removes the need for a separate physics world, reduces overhead, and makes the source code much more approachable and easy to inspect for Bevy users.

In part thanks to Bevy’s modular architecture and the ECS, Bevy XPBD is also highly composable, as it consists of several independent plugins and provides lots of options for configuration and extensions, from custom schedules and plugins to custom joints and constraints.

In terms of the physics implementation, Rapier uses an impulse/velocity based solver, while Bevy XPBD uses Extended Position Based Dynamics. On paper, XPBD should be more stable and robust, but it hasn’t been widely adopted in mainstream usage yet.

One of the biggest disadvantages of Bevy XPBD is that it is still very young, so it can have lots of bugs, some missing features, and fewer community resources and third party crates. However, it is growing quite rapidly, and it is already pretty close to feature-parity with Rapier.

At the end of the day, both engines are very solid options. If you are looking for a more mature and tested physics integration, bevy_rapier is the better choice, but if you prefer an engine with less overhead and a more native Bevy integration, consider using Bevy XPBD. Their core APIs are also quite similar, so switching between them should be straightforward.

Why is nothing happening?

Make sure you have added the PhysicsPlugins plugin group and you have given your rigid bodies a RigidBody component. See the getting started section.

Why is everything moving so slowly?

If your application is in 2D, you might be using pixels as length units. This will require you to use larger velocities and forces than you would in 3D. Make sure you set Gravity to some larger value as well, because its magnitude is 9.81 by default, which is tiny in pixels.

Bevy XPBD doesn’t have a “physics scale” yet, but it will most likely be added in the future so that it’s possible to define some kind of pixels per meter configuration.

Why did my rigid body suddenly vanish?

Make sure to give your rigid bodies some mass, either by adding a Collider or a MassPropertiesBundle. If your bodies don’t have any mass, any physical interaction is likely to instantly give them infinite velocity.

Bevy XPBD should automatically print warnings when it detects bodies with an invalid mass or inertia.

Why is performance so bad?

Make sure you are building your project in release mode using cargo build --release.

You can further optimize builds by setting the number of codegen units in your Cargo.toml to 1, although this will also increase build times.

[profile.release]
codegen-units = 1

Note that Bevy XPBD simply isn’t very optimized yet, and it mostly runs on a single thread for now. This will be addressed in future releases.

Can the engine be used on servers?

Yes! Networking often requires running the simulation in a specific schedule, and in Bevy XPBD you can set the schedule that runs physics and configure the timestep to whatever you want.

One configuration is to run the client in FixedUpdate, and to use PhysicsTimestep::FixedOnce on both the server and the client to make sure the physics simulation is only advanced by one step each time the schedule runs.

Note that while Bevy XPBD should be locally deterministic, it can produce slightly different results on different machines.

Something else?

Physics engines are very large and Bevy XPBD is very young, so stability issues and bugs are to be expected.

If you encounter issues, please consider first taking a look at the issues on GitHub and open a new issue if there already isn’t one regarding your problem.

You can also come and say hello on the Bevy Discord server. There you can find a bevy_xpbd thread on the crate-help channel where you can ask questions.

What is XPBD?

XPBD or Extended Position Based Dynamics is a physics simulation method that extends the traditional PBD to be more physically accurate and less dependent on time step size and iteration count.

Unlike force or impulse based physics simulation methods, XPBD mostly operates at the position-level, which can produce more stable and reliable results, while allowing straightforward coupling of rigid bodies, soft bodies and fluids.

Simulation loop

At a high level, XPBD consists of a broad phase followed by a substepping loop that handles position integration, constraint solving, velocity updates, and a velocity solver that handles dynamic friction and restitution.

It looks roughly like this:

while simulating:
    // Substep size
    h = ∆t / substep_count

    // Broad phase
    collect_collision_pairs()

    for substep_count:
        // Integrate
        for n particles and bodies:
            // Integrate position
            x_prev = x
            v = v + h * f_ext / m
            x = x + h * v

            // Integrate rotation
            q_prev = q
            ω = ω + h * I^-1 * (τ_ext - (ω x (I * ω)))
            q = q + h * 0.5 * [ω_x, ω_y, ω_z, 0] * q
            q = q / |q|

        // Solve constraints (1 iteration and many substeps recommended)
        for iteration_count:
            solve_constraints(particles and bodies)

        // Update velocities
        for n particles and bodies:
            v = (x - x_prev) / h
            ∆q = q * q_prev^-1
            ω = 2 * [∆q_x, ∆q_y, ∆q_z] / h
            ω = ∆q_w >= 0 ? ω : -ω

        // Solve velocity constraints (dynamic friction and restitution)
        solve_velocities(particles and bodies)

where h is the substep size, q is the rotation as a quaternion, ω is the angular velocity, I is the angular inertia tensor and τ is the external torque.

In Bevy XPBD, the simulation loop is handled by various plugins. The PhysicsSetupPlugin sets up the Bevy schedules12 and sets345, the BroadPhasePlugin manages the broad phase, the IntegratorPlugin handles XPBD integration, and so on. You can find all of the plugins and their responsibilities here.

See also

XPBD integration step

Constraints and how to create them

Schedules and sets used for the simulation loop

Learning resources

If you want to learn more about XPBD, I recommend taking a look at some of the papers. Especially the first one from 2020 was used heavily for the simulation loop and constraints in Bevy XPBD.

XPBD: Müller M, Macklin M, Chentanez N, Jeschke S, Kim T. 2020. Detailed Rigid Body Simulation with Extended Position Based Dynamics.

XPBD: Macklin M, Müller M, Chentanez N. 2016. XPBD: Position-Based Simulation of Compliant Constrained Dynamics.

The papers are quite academic, so you might instead prefer some videos and articles. The first one by Ten Minute Physics (Matthias Müller, one of the XPBD researchers) is great for understanding how XPBD differs from other simulation methods and how the constraints work.

Video: Ten Minute Physics. 2022. Getting ready to simulate the world with XPBD.

Notes: Nolan Tait. Bevy Physics: XPBD.

Tutorial series: Johan Helsing. Tutorial: Making a physics engine with Bevy. (inspired this project)

License

Bevy XPBD is free and open source. All code in the Bevy XPBD repository is dual-licensed under either:

MIT License (LICENSE-MIT or http://opensource.org/licenses/MIT)

Apache License, Version 2.0 (LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0)

at your option.

PhysicsSchedule ↩

SubstepSchedule ↩

PhysicsSet ↩

PhysicsStepSet ↩

SubstepSet ↩

Re-exports

pub extern crate parry2d_f64 as parry;

pub use prelude::setup::pause;

pub use prelude::setup::resume;

Modules

collision

Collision events, contact data and helpers.

components

Components used for rigid bodies, colliders and mass properties.

constraints

Contains constraints used by the solver.

math

Math types and traits used in the crate. Most of the math types are feature-dependent, so they will be different for 2d/3d and f32/f64.

plugins

The building blocks of Bevy XPBD.

prelude

Re-exports common components, bundles, resources, plugins and types.

resources

Resources used in the simulation.

Structs

PhysicsSchedule

Responsible for advancing the physics simulation. This is run in PhysicsSet::StepSimulation.

SubstepSchedule

The substepping schedule that runs in PhysicsStepSet::Substeps. The number of substeps per physics step is configured through the SubstepCount resource.

Enums

PhysicsSet

High-level system sets for the main phases of the physics engine. You can use these to schedule your own systems before or after physics is run without having to worry about implementation details.

PhysicsStepSet

System sets for the main steps in the physics simulation loop. These are typically run in the PhysicsSchedule.

SubstepSet

System sets for the the steps in the inner substepping loop. These are typically run in the SubstepSchedule.