Module bevy_xpbd_2d::components

Commonly used components.

Structs

AccumulatedTranslation

Translation accumulated during a sub-step.

AngularDamping

Automatically slows down a dynamic rigid body, decreasing its angular velocity each frame. This can be used to simulate air resistance.

AngularVelocity

The angular velocity of a rigid body in radians per second. Positive values will result in counterclockwise rotation.

CenterOfMass

The local center of mass of a body.

ColliderDensity

The density of a Collider, 1.0 by default. This is used for computing the ColliderMassProperties for each collider.

ColliderMassProperties

An automatically added component that contains the read-only mass properties of a Collider. The density used for computing the mass properties can be configured using the ColliderDensity component.

CollisionLayers

Defines the collision layers of a collider using memberships and filters.

Dominance

Dominance allows dynamic rigid bodies to dominate each other during physical interactions.

ExternalAngularImpulse

An external angular impulse applied instantly to a dynamic rigid body.

ExternalForce

An external force applied continuously to a dynamic rigid body.

ExternalImpulse

An external impulse applied instantly to a dynamic rigid body.

ExternalTorque

An external torque applied continuously to a dynamic rigid body.

Friction

Controls how strongly the material of an entity prevents relative tangential movement at contact points.

GravityScale

Controls how gravity affects a specific rigid body.

Inertia

The moment of inertia of a body. This represents the torque needed for a desired angular acceleration.

InverseInertia

The inverse moment of inertia of the body. This represents the inverse of the torque needed for a desired angular acceleration.

InverseMass

The inverse mass of a body.

LayerMask

A bitmask for layers.

LinearDamping

Automatically slows down a dynamic rigid body, decreasing its linear velocity each frame. This can be used to simulate air resistance.

LinearVelocity

The linear velocity of a rigid body.

LockedAxes

A component that specifies which translational and rotational axes of a rigid body are locked.

Mass

The mass of a body.

MassPropertiesBundle

A bundle containing mass properties.

MassPropertiesQuery

MassPropertiesQueryItem

Automatically generated [WorldQuery] item type for MassPropertiesQuery, returned when iterating over query results.

MassPropertiesQueryReadOnly

Automatically generated [WorldQuery] type for a read-only variant of MassPropertiesQuery.

MassPropertiesQueryReadOnlyItem

Automatically generated [WorldQuery] item type for MassPropertiesQueryReadOnly, returned when iterating over query results.

Position

The global position of a rigid body or a collider.

PreviousPosition

The position of a rigid body at the start of a substep.

PreviousRotation

The previous rotation of a body. See Rotation.

Restitution

A component for the coefficient of restitution. This controls how bouncy a rigid body is.

RigidBodyQuery

A WorldQuery to make querying and modifying rigid bodies more convenient.

RigidBodyQueryItem

Automatically generated [WorldQuery] item type for RigidBodyQuery, returned when iterating over query results.

RigidBodyQueryReadOnly

Automatically generated [WorldQuery] type for a read-only variant of RigidBodyQuery.

RigidBodyQueryReadOnlyItem

Automatically generated [WorldQuery] item type for RigidBodyQueryReadOnly, returned when iterating over query results.

Rotation

The global rotation of a rigid body or a collider.

Sleeping

Indicates that a rigid body is not simulated by the physics engine until woken up again. This is done to improve performance and to help prevent small jitter that is typically present in collisions.

SleepingDisabled

Indicates that the body can not be deactivated by the physics engine. See Sleeping for information about sleeping.

TimeSleeping

How long the velocity of the body has been below the SleepingThreshold, i.e. how long the body has been able to sleep.

Enums

CoefficientCombine

Determines how coefficients are combined for Restitution and Friction. The default is Average.

RigidBody

A non-deformable body used for the simulation of most physics objects.

Traits

PhysicsLayer

A layer used for determining which entities should interact with each other. Physics layers are used heavily by CollisionLayers.