Module bevy_xpbd_2d::plugins::collision

Collision detection is used to detect and compute intersections between Colliders.

In bevy_xpbd, collision detection is split into three plugins:

BroadPhasePlugin: Collects pairs of potentially colliding entities into BroadCollisionPairs.

NarrowPhasePlugin: Computes contacts for broad phase collision pairs and adds them to Collisions.

ContactReportingPlugin (optional): Sends collision events and updates CollidingEntities based on Collisions.

Spatial queries are handled by the SpatialQueryPlugin.

You can also find several utility methods for computing contacts in contact_query.

Modules

broad_phase

Collects pairs of potentially colliding entities into BroadCollisionPairs using AABB intersection checks.

collider_backend

Handles generic collider backend logic, like initializing colliders and AABBs and updating related components.

contact_query

Contact queries compute information about contacts between two Colliders.

contact_reporting

Sends collision events and updates CollidingEntities.

narrow_phase

Computes contacts between entities.

Structs

Collider

A collider used for detecting collisions and generating contacts.

ColliderAabb

The Axis-Aligned Bounding Box of a collider.

ColliderParent

A component that stores the Entity ID of the RigidBody that a Collider is attached to.

ColliderTransform

The transform of a collider relative to the rigid body it’s attached to. This is in the local space of the body, not the collider itself.

CollidingEntities

A component that stores the entities that are colliding with an entity.

Collisions

A resource that stores all collision pairs.

ContactData

Data related to a contact between two bodies.

ContactManifold

A contact manifold between two colliders, containing a set of contact points. Each contact in a manifold shares the same contact normal.

Contacts

All contacts between two colliders.

Sensor

A component that marks a Collider as a sensor, also known as a trigger.

SingleContact

Data related to a single contact between two bodies.

Traits

AnyCollider

A trait that generalizes over colliders. Implementing this trait allows colliders to be used with the physics engine.

IntoCollider

A trait for creating colliders from other types.

ScalableCollider

A trait for colliders that support scaling.

Type Aliases

TriMeshFlags

Flags used for the preprocessing of a triangle mesh collider.

VHACDParameters

Parameters controlling the VHACD convex decomposition algorithm.